fix: 유저 정보 조회시 name이 잘못 담겨오는 문제 해결
